requirements.txt
# Hyperliquid Python SDK
hyperliquid-python-sdk>=0.4.0

# Ethereum account management
eth-account>=0.11.0
web3>=6.0.0

# Environment variable management
python-dotenv>=1.0.0

# Terminal output formatting
tabulate>=0.9.0

# Additional dependencies
requests>=2.31.0
